Se hace 1 request por cada serie destacada a buscar.
    refs:#168
