+++ src/components/mainpage/MainPage.tsx
@@ -1,13 +1,10 @@
 import * as React from 'react';
 import {connect} from 'react-redux';
 import {withRouter} from 'react-router-dom';
-
-import SeriesHero from '../style/Hero/SeriesHero';
-
-import QueryParams from "../../api/QueryParams";
 import {ISerieApi} from '../../api/SerieApi';
 import {IStore} from '../../store/initialState';
 import SearchBox from '../common/searchbox/SearchBox';
+import SeriesHero from '../style/Hero/SeriesHero';
 import Featured from './featured/Featured';
 
 interface IMainPageProps {
@@ -17,6 +14,7 @@ interface IMainPageProps {
     featured: string[];
 }
 
+
 export class MainPage extends React.Component<IMainPageProps, any> {
 
     constructor(props: any, context: any) {
@@ -24,15 +22,6 @@ export class MainPage extends React.Component<IMainPageProps, any> {
 
         this.redirectToSearchPage = this.redirectToSearchPage.bind(this);
         this.redirectToViewPage = this.redirectToViewPage.bind(this);
-
-        this.state = {
-            featuredLoaded: false,
-            featuredSeries: []
-        }
-    }
-
-    public componentWillMount() {
-        this.fetchFeaturedSeries();
     }
 
     public redirectToSearchPage(searchTerm: string) {
@@ -47,18 +36,10 @@ export class MainPage extends React.Component<IMainPageProps, any> {
         return (
             <section id="home">
                 <SeriesHero searchBox={<SearchBox seriesApi={this.props.seriesApi} onSearch={this.redirectToSearchPage} onSelect={this.redirectToViewPage}/>}/>
-                <Featured featured={this.state.featuredSeries} featuredLoaded={this.state.featuredLoaded}/>
+                <Featured featured={this.props.featured} seriesApi={this.props.seriesApi} />
             </section>
         );
     }
-
-    private fetchFeaturedSeries() {
-        const params = new QueryParams(this.props.featured);
-        this.props.seriesApi.fetchMetadata(params).then(featuredSeries => {
-            this.setState({featuredLoaded: true, featuredSeries});
-        })
-    }
-
 }
