+++ src/components/mainpage/featured/Featured.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
-
-import { ISerie } from '../../../api/Serie';
-
-import LoadingSpinner from "../../common/LoadingSpinner";
+import QueryParams from "../../../api/QueryParams";
+import {ISerie} from '../../../api/Serie';
+import {ISerieApi} from "../../../api/SerieApi";
 import FeaturedSerieCard from '../../style/Card/Serie/FeaturedSerieCard';
 import Row from '../../style/Common/Row';
 import FeaturedContainer from '../../style/Featured/FeaturedContainer';
@@ -10,25 +9,47 @@ import FeaturedTitle from '../../style/Featured/FeaturedTitle';
 
 
 interface IFeaturedProps {
-    featured: ISerie[];
-    featuredLoaded: boolean;
+    featured: string[];
+    seriesApi: ISerieApi;
 }
 
-class Featured extends React.Component<IFeaturedProps, any> {
+
+export default class Featured extends React.Component<IFeaturedProps, any> {
+
+    public constructor(props: IFeaturedProps) {
+        super(props);
+
+        this.state = {
+            featuredSeries: []
+        }
+    }
+
+    public componentWillMount() {
+        this.fetchFeaturedSeries();
+    }
 
     public render() {
         return (
             <FeaturedContainer>
                 <FeaturedTitle>Series Destacadas:</FeaturedTitle>
 
-                {(!this.props.featuredLoaded) ? <LoadingSpinner /> : null }
-
                 <Row>
-                    {this.props.featured.map((serie: ISerie) => <FeaturedSerieCard key={serie.id} serie={serie} />)}
+                    {this.props.featured.map((id: string) => {
+                        const serie = this.state.featuredSeries.find((s: ISerie) => s.id === id);
+
+                        return serie ? <FeaturedSerieCard key={serie.id} serie={serie} /> : null
+                    })}
                 </Row>
             </FeaturedContainer>
         );
     }
-}
 
-export default Featured;
+    private fetchFeaturedSeries() {
+        this.props.featured.forEach((id: string) => {
+            this.props.seriesApi.fetchSeries(new QueryParams([id])).then(featuredSeries => {
+                this.setState({featuredSeries: Array.from(this.state.featuredSeries).concat(featuredSeries)});
+            })
+        });
+    }
+
+}
